refactor: move llm playground logic to shared package (#1476)
