+++ packages/shared/src/index.ts
@@ -3,4 +3,7 @@ export { type DB } from "../prisma/generated/types";
 
 export * from "./auth/auth";
 
+export * from "./server/llm/types";
+export * from "./server/llm/fetchLLMCompletion";
+
 export { ModelUsageUnit };
