+++ packages/shared/tsconfig.json
@@ -1,11 +1,12 @@
 {
   "extends": "@repo/typescript-config/base.json",
   "compilerOptions": {
-    "moduleResolution": "Node",
-    "module": "commonjs",
-    "lib": ["ES2015"],
+    "moduleResolution": "NodeNext",
+    "module": "NodeNext",
+    "lib": ["ES2020"],
     "outDir": "./dist",
-    "types": ["node"]
+    "types": ["node"],
+    "target": "ES2020"
   },
   "include": ["."],
   "exclude": ["node_modules", "dist"]
