+++ web/src/features/playground/client/components/PromptVariableComponent.tsx
@@ -3,7 +3,7 @@ import { usePlaygroundContext } from "../context";
 import { Input } from "@/src/components/ui/input";
 import { CheckCircle2, Circle, Trash2Icon } from "lucide-react";
 import { Button } from "@/src/components/ui/button";
-import { type PromptVariable } from "@/src/features/playground/types";
+import { type PromptVariable } from "@langfuse/shared";
 
 export const PromptVariableComponent: React.FC<{
   promptVariable: PromptVariable;
